Restructure file structure (#18)
